OTUS homework 002: Lucky tickets -- LuckyTicket.java function type 'findAmount' changed to 'double' due to restrictions of Long; corresponding changes in Tester.java.

+++ HomeWork001/src/LuckyTicket.java
@@ -12,7 +12,7 @@ public class LuckyTicket {
         }
     }
 
-    long findAmount(int len_num) {
+    double findAmount(int len_num) {
         if (len_num < 1)
             return 0;
         if (len_num == 1)
